fix findbugs high priority warnings
